Fix apocalyptically bad checksum calcs
First off, you can't just bitshift them a couple of places, because
then, among other things, 2 4-byte packets differing only in the low
half of the third byte would often (due to the in-packet checksum) come
out the same.

Worse still, instead of a 32-bit checksum, I had a 16-bit checksum,
because I was getting integer-math'ed since I forgot to cast it as an
unsigned long.